# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "utopia API",
        "description": "API documentation for utopia ",
        "contact": {
            "email": "info@utopia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gestion.utopia.madrid/api/",
            "description": "utopia API"
        }
    ],
    "paths": {
...
        "/contactos": {
            "post": {
                "tags": [
                    "contactos"
                ],
                "summary": "Crear contacto",
                "operationId": "2b7ef1175492748f2e2d5cbf94d24a62",
                "requestBody": {
                    "description": "Post contacto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "identidad"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456789"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_@_.com"
                                    },
                                    "identidad": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/conatctos/{id}": {
            "get": {
                "tags": [
                    "contactos"
                ],
                "summary": "Mostrar contacto ",
                "operationId": "3eca4e58bbef14a0a6aef2bb25d75177",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta de contacto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/contactos/{id}": {
            "put": {
                "tags": [
                    "contactos"
                ],
                "summary": "Modificar contacto",
                "operationId": "9d2a330a90e4a69e815474b5befc592c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro para editar contacto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "put contacto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_@_.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456789"
                                    },
                                    "identidad": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "contactos"
                ],
                "summary": "Eliminar contacto",
                "operationId": "522ff9d198f2edeca087cc8e774e58a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro para editar contacto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}